requirements.txt
azure-cognitiveservices-speech==1.38.0
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.29.0
gradio==4.26.0